Change API to return result ordering
