+++ datafusion/physical-expr/src/window/built_in.rs
@@ -23,9 +23,13 @@ use std::sync::Arc;
 
 use super::BuiltInWindowFunctionExpr;
 use super::WindowExpr;
+use crate::equivalence::OrderingEquivalenceBuilder;
+use crate::utils::{convert_to_expr, get_indices_of_matching_exprs};
 use crate::window::window_expr::WindowFn;
 use crate::window::{PartitionBatches, PartitionWindowAggStates, WindowState};
-use crate::{expressions::PhysicalSortExpr, reverse_order_bys, PhysicalExpr};
+use crate::{
+    expressions::PhysicalSortExpr, reverse_order_bys, EquivalenceProperties, PhysicalExpr,
+};
 use arrow::array::{new_empty_array, ArrayRef};
 use arrow::compute::SortOptions;
 use arrow::datatypes::Field;
@@ -64,6 +68,45 @@ impl BuiltInWindowExpr {
     /// Get BuiltInWindowFunction expr of BuiltInWindowExpr
     pub fn get_built_in_func_expr(&self) -> &Arc<dyn BuiltInWindowFunctionExpr> {
         &self.expr
+    }
+
+    /// Adds any equivalent orderings generated by this expression
+    /// to `builder`.
+    ///
+    /// The default implementation does nothing
+    /// If `prefix` is given newly introduced ordering is prepended by `prefix`.
+    pub fn add_equal_orderings<F: FnOnce() -> EquivalenceProperties>(
+        &self,
+        builder: &mut OrderingEquivalenceBuilder,
+        equal_properties: F,
+    ) {
+        let schema = builder.schema();
+        if let Some(fn_res_ordering) = self.expr.get_result_ordering(schema) {
+            // if expr contains partition by, ROW_NUMBER cannot introduce global ordering.
+            if self.partition_by.is_empty() {
+                // Function result ordering is global
+                builder.add_equal_conditions(vec![fn_res_ordering]);
+            } else {
+                let existing_ordering = builder.existing_ordering();
+                let existing_ordering_exprs = convert_to_expr(existing_ordering);
+                // indices of the partition by expressions among input ordering expressions
+                let pb_indices = get_indices_of_matching_exprs(
+                    &self.partition_by,
+                    &existing_ordering_exprs,
+                    equal_properties,
+                );
+                // Existing ordering should match exactly with partition by expressions
+                // there shouldn't be missing entry or additional entry in the existing ordering
+                // otherwise prefix wouldn't work
+                if pb_indices.len() == self.partition_by.len()
+                    && pb_indices.len() == existing_ordering.len()
+                {
+                    let mut new_ordering = existing_ordering.to_vec();
+                    new_ordering.push(fn_res_ordering);
+                    builder.add_equal_conditions(new_ordering);
+                }
+            }
+        }
     }
 }
