+++ datafusion/physical-expr/src/window/built_in_window_function_expr.rs
@@ -15,11 +15,11 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use crate::equivalence::OrderingEquivalenceBuilder;
-use crate::{LexOrdering, PhysicalExpr};
+use crate::{PhysicalExpr, PhysicalSortExpr};
 use arrow::array::ArrayRef;
 use arrow::datatypes::Field;
 use arrow::record_batch::RecordBatch;
+use arrow_schema::SchemaRef;
 use datafusion_common::Result;
 use datafusion_expr::PartitionEvaluator;
 use std::any::Any;
@@ -80,15 +80,10 @@ pub trait BuiltInWindowFunctionExpr: Send + Sync + std::fmt::Debug {
         None
     }
 
-    /// Adds any equivalent orderings generated by this expression
-    /// to `builder`.
-    ///
-    /// The default implementation does nothing
-    /// If `prefix` is given newly introduced ordering is prepended by `prefix`.
-    fn add_equal_orderings(
-        &self,
-        _builder: &mut OrderingEquivalenceBuilder,
-        _prefix: Option<LexOrdering>,
-    ) {
+    /// Window function result may have ordering. If this is the case, this method should return
+    /// ordering of the window function (This information is used to update ordering equivalence)
+    /// By default result doesn't introduce ordering e.g None.
+    fn get_result_ordering(&self, _schema: &SchemaRef) -> Option<PhysicalSortExpr> {
+        None
     }
 }
